+++ src/lib/types.ts
@@ -546,7 +546,7 @@ type ContactPreferences = "TELEPHONE" | "EMAIL" | "AUTRE";
 export interface Orientation {
   // Tous les champs de l'étape 1 pouvant être optionnels
   // on est contraint de se baser sur un booléen pour s'assurer que l'étape 1 a bien été visionnée
-  firstStepView: boolean;
+  firstStepDone: boolean;
 
   situation: string[];
   situationOther: string;
