+++ src/routes/(modeles-services)/services/[slug]/orienter/demande/+page.ts
@@ -1,5 +1,14 @@
+import { goto } from "$app/navigation";
+import { get } from "svelte/store";
+import { orientation } from "../store";
+
 export const load = async ({ parent }) => {
   const data = await parent();
+
+  if (!get(orientation).firstStepDone) {
+    goto(`/services/${data.service.slug}/orienter`);
+    return {};
+  }
 
   return {
     ...data,
